feat: combine all helpers into one file for reduce library size; docs: add JSDOC for most functions in helpers
