+++ lib/kinka.js
@@ -6,9 +6,9 @@
  *
  * @flow
  */
-var baseHelpers = require('./helpers/base')
-var requestHelpers = require('./helpers/request')
-var isUndefined = baseHelpers.isUndefined
+
+var helpers = require('./helpers')
+var isUndefined = helpers.isUndefined
 
 var DEFAULT_CHARSET = 'utf-8'
 
@@ -37,12 +37,12 @@ function createMethodHandler(methodName, hasBody) {
 }
 
 var kinkaPrototype = {
-  abort: requestHelpers.abortRequest,
+  abort: helpers.abortRequest,
   all: Promise.all,
   create: function(instanceOptions) {
     return new Kinka(instanceOptions)
   },
-  custom: requestHelpers.createRequest,
+  custom: helpers.createRequest,
   delete: createMethodHandler('delete'),
   get: createMethodHandler('get'),
   head: createMethodHandler('head'),
